Added Windows build support (#25)
* midistar can now be built on Windows

* Added scripts to make / run in Debug / Release modes on Windows and OSX/Linux

* Build setup now automatically builds midistar

* Updated SFML version to get Ubuntu build error bug fix

+++ src/DefaultGameObjectFactory.cpp
@@ -93,11 +93,11 @@ GameObject* DefaultGameObjectFactory::CreateInstrumentNote(int note) {
     double y = Config::GetInstance().GetScreenHeight() - INSTRUMENT_HEIGHT -
         (Config::GetInstance().GetScreenHeight() * INSTRUMENT_HOVER_PERCENTAGE);
     sf::RectangleShape* rect = new sf::RectangleShape{{static_cast<float>(
-            note_width_), INSTRUMENT_HEIGHT}};;
+            note_width_), static_cast<float>(INSTRUMENT_HEIGHT)}};;
     rect->setFillColor(sf::Color::Red);
 
     // Create GameObject
-    auto ins_note = new GameObject{rect, x, y, note_width_, INSTRUMENT_HEIGHT};
+    auto ins_note = new GameObject{rect, x, y, note_width_, static_cast<float>(INSTRUMENT_HEIGHT)};
 
     // Get key binding
     sf::Keyboard::Key key;
